backend/requirements.txt
fastapi
uvicorn
pymongo
requests
bcrypt
pyjwt
apscheduler
python-multipart
passlib
cryptography
dnspython
email-validator
pydantic
python-dotenv
slowapi
